backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.2
supabase>=2.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
email-validator>=2.1.0 